requirements.txt
django==2.1.5
python-decouple==3.1
PyJWT==1.7.1
